+++ waves/_tests/test_one_at_a_time.py
@@ -186,8 +186,8 @@ class TestOneAtATime:
             # Ordered by md5 hash during Xarray merge operation. New tests must verify hash ordering.
             numpy.array(
                 [
-                    [1, True, "a"],
                     [2, True, "a"],
+                    [1, True, "a"],
                 ],
                 dtype=object,
             ),
@@ -217,7 +217,7 @@ class TestOneAtATime:
             None,
             "yaml",
             2,
-            [call("parameter_1: 1\n"), call("parameter_1: 2\n")],
+            [call("parameter_1: 2\n"), call("parameter_1: 1\n")],
         ),
         "two parameter yaml": (
             {"parameter_1": [1, 2], "parameter_2": ["a", "b"]},
@@ -227,8 +227,8 @@ class TestOneAtATime:
             3,
             [
                 call("parameter_1: 1\nparameter_2: a\n"),
-                call("parameter_1: 2\nparameter_2: a\n"),
                 call("parameter_1: 1\nparameter_2: b\n"),
+                call("parameter_1: 2\nparameter_2: a\n"),
             ],
         ),
         "two parameter yaml: floats and ints": (
@@ -238,9 +238,9 @@ class TestOneAtATime:
             "yaml",
             3,
             [
+                call("parameter_1: 1\nparameter_2: 4.0\n"),
                 call("parameter_1: 1\nparameter_2: 3.0\n"),
                 call("parameter_1: 2\nparameter_2: 3.0\n"),
-                call("parameter_1: 1\nparameter_2: 4.0\n"),
             ],
         ),
         "two parameter yaml: bools and ints": (
@@ -250,8 +250,8 @@ class TestOneAtATime:
             "yaml",
             3,
             [
-                call("parameter_1: 1\nparameter_2: true\n"),
                 call("parameter_1: 2\nparameter_2: true\n"),
+                call("parameter_1: 1\nparameter_2: true\n"),
                 call("parameter_1: 1\nparameter_2: false\n"),
             ],
         ),
@@ -261,7 +261,7 @@ class TestOneAtATime:
             "parameter_study.yaml",
             "yaml",
             1,
-            [call("parameter_set0:\n  parameter_1: 1\nparameter_set1:\n  parameter_1: 2\n")],
+            [call("parameter_set0:\n  parameter_1: 2\nparameter_set1:\n  parameter_1: 1\n")],
         ),
         "two parameter one file yaml": (
             {"parameter_1": [1, 2], "parameter_2": ["a", "b"]},
@@ -272,8 +272,8 @@ class TestOneAtATime:
             [
                 call(
                     "parameter_set0:\n  parameter_1: 1\n  parameter_2: a\n"
-                    "parameter_set1:\n  parameter_1: 2\n  parameter_2: a\n"
-                    "parameter_set2:\n  parameter_1: 1\n  parameter_2: b\n"
+                    "parameter_set1:\n  parameter_1: 1\n  parameter_2: b\n"
+                    "parameter_set2:\n  parameter_1: 2\n  parameter_2: a\n"
                 )
             ],
         ),
@@ -285,8 +285,8 @@ class TestOneAtATime:
             1,
             [
                 call(
-                    "parameter_set0:\n  parameter_1: 1\n  parameter_2: true\n"
-                    "parameter_set1:\n  parameter_1: 2\n  parameter_2: true\n"
+                    "parameter_set0:\n  parameter_1: 2\n  parameter_2: true\n"
+                    "parameter_set1:\n  parameter_1: 1\n  parameter_2: true\n"
                     "parameter_set2:\n  parameter_1: 1\n  parameter_2: false\n"
                 )
             ],
@@ -327,11 +327,11 @@ class TestOneAtATime:
         ),
         "floats": (
             {"floats": [10.0, 20.0]},
-            {"parameter_set0": {"floats": 10.0}, "parameter_set1": {"floats": 20.0}},
+            {"parameter_set0": {"floats": 20.0}, "parameter_set1": {"floats": 10.0}},
         ),
         "strings": (
             {"strings": ["a", "b"]},
-            {"parameter_set0": {"strings": "a"}, "parameter_set1": {"strings": "b"}},
+            {"parameter_set0": {"strings": "b"}, "parameter_set1": {"strings": "a"}},
         ),
         "bools": (
             {"bools": [False, True]},